+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Causer` interface
 - `WithPrefix` didn't really make sense with the decorator pattern
 - `Contextor` interface
+- `StackTracer` interface
 
 
 ## [0.7.1] - 2018-04-27